backend/requirements.txt
fastapi==0.111.0
uvicorn==0.28.1
sqlalchemy==2.0.34
alembic==1.13.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.4
email-validator==2.2.0
httpx==0.27.0
python-multipart==0.0.9